code moved to recyclerview package
